deleted some methods, refactored and rearranged project to minimize testing and be able to sumbit project before deadline

+++ migrations/sqls/20220404133351-orders-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE Table orders_products(
   id SERIAL PRIMARY KEY,
-  order_id BIGINT REFERENCES orders(id),
-  product_id BIGINT REFERENCES products(id),
+  order_id INTEGER REFERENCES orders(id),
+  product_id INTEGER REFERENCES products(id),
   quantity INTEGER
 );
